+++ src/main/java/com/kruemblegard/event/CephalariZombieMountLinkEvents.java
@@ -9,6 +9,7 @@ import net.minecraft.world.entity.monster.Husk;
 import net.minecraft.world.entity.monster.Zombie;
 import net.minecraft.world.entity.monster.ZombieVillager;
 import net.minecraft.world.entity.monster.ZombifiedPiglin;
+import net.minecraft.world.entity.monster.Zoglin;
 
 import net.minecraftforge.event.entity.living.LivingHealEvent;
 import net.minecraftforge.event.entity.living.LivingHurtEvent;
@@ -30,7 +31,7 @@ public final class CephalariZombieMountLinkEvents {
             return;
         }
 
-        if (!(mob instanceof Zombie || mob instanceof Husk || mob instanceof Drowned || mob instanceof ZombieVillager || mob instanceof ZombifiedPiglin)) {
+        if (!(mob instanceof Zombie || mob instanceof Husk || mob instanceof Drowned || mob instanceof ZombieVillager || mob instanceof ZombifiedPiglin || mob instanceof Zoglin)) {
             return;
         }
 
@@ -56,7 +57,7 @@ public final class CephalariZombieMountLinkEvents {
             return;
         }
 
-        if (!(mob instanceof Zombie || mob instanceof Husk || mob instanceof Drowned || mob instanceof ZombieVillager || mob instanceof ZombifiedPiglin)) {
+        if (!(mob instanceof Zombie || mob instanceof Husk || mob instanceof Drowned || mob instanceof ZombieVillager || mob instanceof ZombifiedPiglin || mob instanceof Zoglin)) {
             return;
         }